fixing data type errors

+++ 2_assembly/config_dycd_facilities_afterschoolprograms.sql
@@ -70,7 +70,7 @@ SELECT
 	-- borough
 	initcap(BOROUGH_COMMUNITY),
 	-- zipcode
-	Postcode,
+	Postcode::integer,
 	-- bbl
 	BBL,
 	-- bin
